+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Changelog for Cldr v0.0.17 February 21st, 2017
+
+### Enhancements
+
+* `Cldr.version/0` returns the version of the CLDR repository in use.  This is now automatically deterived from the library data.
+
 ## Changelog for Cldr v0.0.17 February 20th, 2017
 
 ### Bug fixes